move EldersChangeBuilder from chain to consensus::network_event
